+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0-alpha.6
+
+* Fix model and form resolution during client-side re-render when element is not yet in DOM
+* Add fallback to default wrappers when field_type cannot be determined
+
 ## 0.3.0-alpha.5 (2025-07-10)
 
 * Add `al-datetime-input` to properly handle Date values with timezones